emu: pool sizes are uintptr, not int - and the GPU does pay after all
"panic: not enough memory" on a machine with plenty. poolsetsize took a
C int while maxsize and ressize have always been uintptr, so -pheap=2g
arrived negative and -pheap=4g arrived as zero, and the size check
reported it as not enough memory. atoi in poolopt truncated it first for
the same reason. Fixed at both ends, with strtoull, an overflow check
and a g suffix; a bad size now prints usage rather than panicking, and
nothing is modified before the check, so the pool table cannot be left
inconsistent.

This mattered more than it looks. doc/hpc-plan.md, man/3/gpu and
man/1/gpubench all recorded "gpubench panics at mesh 28 and above" as a
fact about the machine, and drew a conclusion from the range it left:
that GPU offload does not pay at any size that fits. A mesh of 28 cubed
needs a few megabytes.

With the ceiling raised:

  mesh       nodes   cpu f64             gpu f32
                     solve   us/iter     solve   us/iter
  24x24x24   15625    5-6ms    147      7-18ms     179
  28x28x28   21952  10-11ms    250     10-23ms     222   <- per matvec
  30x30x30   27000  13-15ms    310     13-25ms     250
  32x32x32   32768  17-18ms    378     15-28ms     288   <- per solve

The GPU wins per matvec from mesh 28 and per solve from 32. Below that
its cost per iteration is flat - 176 to 200us across a sevenfold range -
which is what per-call overhead looks like, while the CPU's grows, which
is what arithmetic looks like.

Two cautions kept with the numbers: the per-solve win is f32 against f64
at 52 iterations against 45, so those are not the same solve; and the
GPU's spread is far wider (15-28ms against 17-18ms), so it wins on the
best of five and not on the worst.

Plan item 2 (device-resident vectors) is un-suspended as a result, with
a note to re-derive its 40% from this table rather than the old one.

Fourth correction to this table. The first three were a wrong baseline,
one-time setup read as per-call cost, and a trend read out of single
noisy samples. This one is a ceiling nobody questioned.

+++ include/pool.h
@@ -81,7 +81,7 @@ extern	void	poolmutable(void*);
 extern	char*	poolname(Pool*);
 extern	int	poolread(char*, int, uintptr);
 extern	void*	poolrealloc(Pool*, void*, uintptr);
-extern	int	poolsetsize(char*, int);
+extern	int	poolsetsize(char*, uintptr);
 extern	void	poolsetcompact(Pool*, void (*)(void*, void*));
 extern	char*	poolaudit(char*(*)(int, Bhdr *));
